move zap to gps, where it is used
