+++ apps/frontend/src/utils/gps.ts
@@ -1,10 +1,17 @@
 import { Waypoint } from '../types';
 import { lerp } from './general';
 
-export const distanceToCoordinates = (
-  path: Waypoint[],
-  totalDistance: number
-) => {
+const zap: Waypoint[] = [
+  { lat: 59.90347154, lon: 10.6590337, distance: 2400 },
+  { lat: 59.88396124, lon: 10.64085992, distance: 600 },
+  { lat: 59.88389387, lon: 10.65213867, distance: 2000 },
+  { lat: 59.86610453, lon: 10.64629091, distance: 2400 },
+  { lat: 59.88561483, lon: 10.6644647, distance: 600 },
+  { lat: 59.8856822, lon: 10.65318595, distance: 2000 },
+];
+
+export const distanceToCoordinates = (totalDistance: number) => {
+  const path = zap;
   // TODO: Memoize this
   const totalPathDistance = path.reduce(
     (sum, waypoint) => sum + waypoint.distance,
